+++ example/tasks/lib/nodus.g.dart
@@ -1,6 +1,6 @@
 // GENERATED FILE. DO NOT EDIT.
 // Source: package:tasks_example/nodus.lock
-// Schema fingerprint: 0218a913de9d5ce875369707cc836f23258fa2ff515b76e11e8e2e17b3785ddd
+// Schema fingerprint: 4339b886eceaee9d143ba7702ddd3f5957ec986062a084e93960ccfedbd4fb6b
 
 export 'package:nodus/nodus_flutter.dart';
 export 'src/generated/nodus.runtime.g.dart';